requirements.txt
# ============================================================================
# requirements.txt
# ============================================================================
"""
google-genai
google-adk
yfinance
python-dotenv
"""